FRES-117:Created banner component
Created a paragraph type for banner with image, title and link field.
Added the same as a field in the basic page CT.
Created a node for the basic page and added the data in paragraphs.
Created a new region named hero and rendered it using page.html.twig.
Created a view of paragraph and placed it in block layout to show on the home page.
Created a components folder, and made a banner component by creating banner.component.yml file and banner.twig file.
Overridden twig template for paragraph with paragraph–banner.html.twig and included the paragraph component with proper variables.
Done styling according to a banner in the banner.css file.